fix: keep sitemap manifest reader server-only
